Split upscale source media hook
